Feature/local storage hydration (#3)
* save progress

* hydrate component from localStorage

* work in share url to gear provider and have it use state instead of router values to generate

* fix errors and update state to handle character class and perks

* clean up GearState useage

* fix state bug

* Fix bugs, update local docker compose, add url hydration from url

* added locking to class layout for share page

* fix a bug

* add detail list to share page

* update quick detail view

* link in header

* hide and show reset kit button

* update pnpm lock

+++ src/components/SingleSlot/SingleSlot.tsx
@@ -15,7 +15,7 @@ import css from './SingleSlot.module.scss'
 import { SlotSize } from '@/common/utils/determineSlotSize'
 import { ExcludeFromUnion } from '@/common/utils/typeHelpers'
 import { MouseEventHandler, useMemo, useState } from 'react'
-import { GearSlots } from '@/providers/GearProvider/types'
+import { GearSlot } from '@/providers/GearProvider/types'
 import { useGear } from '@/providers/GearProvider/GearProvider'
 import { createAbbreviation } from '@/common/utils/createAbbreviation'
 import { getGearScore } from '@/providers/GearProvider/utils'
@@ -23,11 +23,12 @@ import { arrayCamelCaseToCommaSeparatedTitleCase } from '@/common/utils/slugToHu
 
 interface SingleSlotProps {
   item: Item | null
-  slotSlug: GearSlots
+  slotSlug: GearSlot
   slotSize: ExcludeFromUnion<SlotSize, 'largeTwo'>
   itemRarity: string | null
   className?: string
   onLeftClick?: MouseEventHandler<HTMLDivElement>
+  isLocked?: boolean
 }
 
 export default function SingleSlot({
@@ -37,6 +38,7 @@ export default function SingleSlot({
   itemRarity,
   className,
   onLeftClick,
+  isLocked = false,
 }: SingleSlotProps) {
   const [isOpen, setIsOpen] = useState(false)
   const { deleteSlot } = useGear()
@@ -62,7 +64,9 @@ export default function SingleSlot({
   }, [item])
 
   const handleSlotRightCLick = () => {
-    deleteSlot(slotSlug)
+    if (!isLocked) {
+      deleteSlot(slotSlug)
+    }
   }
 
   return (
